+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 -   Entirely removed the `Group` class. This means code utilizing groups will need to be significantly refactored.
 -   Refactored `Scene`s as a consequence of the removal of groups. See documentation for more information.
+-   Added `ignore_cam` option to `GameObject`s. This decides whether the gameobject respects the scene's camera when it draws.
 
 ### Added